--- unity/PackageProject/Assets/Standard Assets/Editor.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: daf2194cf3b6c4b6d85b044ceba60f69
-folderAsset: yes
-timeCreated: 1531339818
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
