--- unity/PackageProject/Assets/Standard Assets/Editor/BugsnagPostProcess.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: bbcdeee05b6934841946e1d7a4b07d3b
-timeCreated: 1531339822
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
